Add day 9 solutions

+++ src/main/java/net/akaritakai/aoc2019/Puzzle07.java
@@ -26,9 +26,9 @@ public class Puzzle07 extends AbstractPuzzle {
 
   @Override
   public String solvePart1() {
-    var maxSignal = Collections2.permutations(List.of(0, 1, 2, 3, 4))
+    var maxSignal = Collections2.permutations(List.of(0L, 1L, 2L, 3L, 4L))
         .stream()
-        .mapToInt(this::runAmplifiers)
+        .mapToLong(this::runAmplifiers)
         .max()
         .orElseThrow(() -> new IllegalStateException("No amplifier outputs returned"));
     return String.valueOf(maxSignal);
@@ -36,9 +36,9 @@ public class Puzzle07 extends AbstractPuzzle {
 
   @Override
   public String solvePart2() {
-    var maxSignal = Collections2.permutations(List.of(5, 6, 7, 8, 9))
+    var maxSignal = Collections2.permutations(List.of(5L, 6L, 7L, 8L, 9L))
         .stream()
-        .mapToInt(this::runAmplifiers)
+        .mapToLong(this::runAmplifiers)
         .max()
         .orElseThrow(() -> new IllegalStateException("No amplifier outputs returned"));
     return String.valueOf(maxSignal);
@@ -48,9 +48,9 @@ public class Puzzle07 extends AbstractPuzzle {
    * Runs a set of amplifiers in a chained feedback loop, applying the phase settings provided to each of them.
    * Returns the last unused output of the last amplifier after all amplifiers halt.
    */
-  private int runAmplifiers(List<Integer> phaseSettings) {
+  private long runAmplifiers(List<Long> phaseSettings) {
     // Create the i/o streams
-    var streams = Stream.generate((Supplier<LinkedBlockingQueue<Integer>>) LinkedBlockingQueue::new)
+    var streams = Stream.generate((Supplier<LinkedBlockingQueue<Long>>) LinkedBlockingQueue::new)
         .limit(phaseSettings.size())
         .collect(Collectors.toList());
     // Create and start the amplifiers
@@ -62,7 +62,7 @@ public class Puzzle07 extends AbstractPuzzle {
           return runAmplifier(phaseSetting, input, output);
         })
         .collect(Collectors.toList());
-    streams.get(0).add(0); // send 0 to the first amplifier
+    streams.get(0).add(0L); // send 0 to the first amplifier
     amplifiers.forEach(sneaked((SneakyConsumer<Thread, Exception>) Thread::join)); // wait for the amplifiers to halt
     return streams.get(0).remove();
   }
@@ -71,7 +71,7 @@ public class Puzzle07 extends AbstractPuzzle {
    * Returns a thread for an amplifier that accepts a phase setting and an input stream. Amplifier outputs will be sent
    * to the output stream.
    */
-  private Thread runAmplifier(int phaseSetting, BlockingQueue<Integer> inputStream, BlockingQueue<Integer> outputStream) {
+  private Thread runAmplifier(long phaseSetting, BlockingQueue<Long> inputStream, BlockingQueue<Long> outputStream) {
     sneaked(() -> inputStream.put(phaseSetting)).run();  // enqueue the phaseSetting to the amplifier's input stream
     var thread = new Thread(() -> {
       var program = getPuzzleInput();
